Add support for multi db setups with a different adapter
I was implementing solid_cable using SQLite as an adapter and came across this issue, where the schema dumper was trying to dump the sequences for the SQLite solid cable schema. Obviously, this isn't possible.

I've modified the code to hook into the `ActiveRecord::ConnectionAdapters::PostgreSQL::SchemaDumper` instead of the generic `ActiveRecord::SchemaDumper` and added the sequences' logic to the `extensions` section (which to me seems like a good place to put this).